backend/requirements.txt
Flask==1.1.1
#flask-cors==3.0.8
Werkzeug==1.0.0
python-dotenv==0.12.0
pytest
pytest-ordering
pandas
elasticsearch===7.9.0
tika==1.19
xlrd
odfpy==1.4.0
Unidecode==1.1.1
sentence-transformers==0.3.8
